- Merge branch 'dev' of ssh.dev.azure.com:v3/TargetAfri/CVC/CVCBackEnd into fixes
- fixed view noa

+++ Bunkering.Access/Services/PaymentService.cs
@@ -186,7 +186,7 @@ namespace Bunkering.Access.Services
                                         payment = new Payment
                                         {
                                             Account = "",
-                                            Amount = total,
+                                            Amount = (decimal)total,
                                             ApplicationId = coq.AppId,
                                             ApplicationTypeId = appType.Id,
                                             COQId = coq.Id,
@@ -267,11 +267,11 @@ namespace Bunkering.Access.Services
                             _response = new ApiResponse { Message = "Debit note does not exist for this Depot", StatusCode = HttpStatusCode.BadRequest };
                         else
                         {
-                            var total = payment.Amount * 0.10;
+                            var total = payment.Amount * 0.10m;
                             var reference = Utils.RefrenceCode();
                             var description = $"Payment for non-payment of Debit note generated for {coq.Depot.Name} after 21 days as regulated";
 
-                            var request = await _elps.GenerateDebitNotePaymentReference($"{_contextAccessor.HttpContext.Request.Scheme}://{_contextAccessor.HttpContext.Request.Host}", total, coq.Application.User.Company.Name, coq.Application.User.Email, coq.Reference, coq.Depot.Name, coq.Application.User.ElpsId, Enum.GetName(typeof(AppTypes), AppTypes.DemandNotice), description);
+                            var request = await _elps.GenerateDebitNotePaymentReference($"{_contextAccessor.HttpContext.Request.Scheme}://{_contextAccessor.HttpContext.Request.Host}", (double)total, coq.Application.User.Company.Name, coq.Application.User.Email, coq.Reference, coq.Depot.Name, coq.Application.User.ElpsId, Enum.GetName(typeof(AppTypes), AppTypes.DemandNotice), description);
                             _logger.LogRequest($"Creation of Demand notice payment for application with reference: {reference}for ({coq.Application.User.Company.Name}) as specified in the Authority regulations", false, directory);
 
                             if (request == null)
